refactor(transactions): extract shared domain layer from Preact components
