+++ packages/domains/transactions/domain/src/TransactionsOverview/analyticsConstants.ts
@@ -1,0 +1,4 @@
+export const TRANSACTION_ANALYTICS_CATEGORY = 'Transaction component' as const;
+export const TRANSACTION_ANALYTICS_SUBCATEGORY_DETAILS = 'Transaction details' as const;
+export const TRANSACTION_ANALYTICS_SUBCATEGORY_INSIGHTS = 'Transactions insights' as const;
+export const TRANSACTION_ANALYTICS_SUBCATEGORY_LIST = 'Transactions list' as const;
