+++ packages/domains/transactions/domain/src/TransactionsOverview/fields.ts
@@ -1,0 +1,6 @@
+export const TRANSACTION_FIELDS = ['createdAt', 'paymentMethod', 'transactionType', 'currency', 'netAmount', 'grossAmount'] as const;
+export type TransactionsTableCols = (typeof TRANSACTION_FIELDS)[number];
+
+export const TRANSACTION_FIELDS_REMAPS: Readonly<Record<string, TransactionsTableCols>> = {
+    amount: 'netAmount',
+} as const;
