+++ packages/domains/transactions/domain/src/index.ts
@@ -1,1 +1,9 @@
 export * from './TransactionDetails/translations';
+export * from './TransactionDetails/types';
+export * from './TransactionDetails/constants';
+export * from './TransactionDetails/utils';
+export * from './TransactionsOverview/analyticsConstants';
+export * from './TransactionsOverview/fields';
+export * from './shared/inputNormalizer';
+export * from './shared/refundLineItemHelpers';
+export * from './shared/transactionNavigator';
