+++ packages/domains/transactions/preact/src/TransactionDetails/utils.ts
@@ -1,17 +1,7 @@
-import { EMPTY_OBJECT } from '@integration-components/utils';
-import { RefundType, TransactionDetails } from './types';
 import { TagVariant } from '@integration-components/ui-components-preact/Tag/types';
+import { TransactionDetails } from './types';
 
-export const getAmountStyleForTransaction = (transaction?: TransactionDetails) => {
-    switch (transaction?.status) {
-        case 'Booked':
-            return 'default';
-        case 'Reversed':
-            return 'error';
-        default:
-            return 'pending';
-    }
-};
+export { getAmountStyleForTransaction, getRefundTypeForTransaction } from '@integration-components/transactions/domain';
 
 export const getTagVariantForTransaction = (transaction?: TransactionDetails) => {
     switch (transaction?.status) {
@@ -21,17 +11,5 @@ export const getTagVariantForTransaction = (transaction?: TransactionDetails) =>
             return TagVariant.ERROR;
         default:
             return TagVariant.DEFAULT;
-    }
-};
-
-export const getRefundTypeForTransaction = (transaction?: TransactionDetails) => {
-    if (transaction?.category === 'Refund') {
-        const { refundType } = transaction.refundMetadata ?? (EMPTY_OBJECT as NonNullable<TransactionDetails['refundMetadata']>);
-        switch (refundType) {
-            case RefundType.FULL:
-                return RefundType.FULL;
-            case RefundType.PARTIAL:
-                return RefundType.PARTIAL;
-        }
     }
 };
